Implement quirks
In some cases there is needed for the update to come with an overlay fs
structure which is intended to be written on top of new rootfs after update.
This overlay FHS needs to be in a directory called quirks and located in the
root of the update files structure.

Use case example: docker shadows resolv.conf so we need to provide it in
another location which would used to copy it in the final rootfs.

Signed-off-by: Andrei Gherzan <andrei@resin.io>