dockerhub.py/tar.py: Replace workspacefile by a file object
As well we need now to test the archive after we unpack it. And this is because
because of the fact that now we are opening the tar file without seek
availability so we can't read first the filenames and then unpack as the data in
tar stream is interleaved (header/data/header/data etc).

Signed-off-by: Andrei Gherzan <andrei@resin.io>