upgrade-2.x: check VERSION_ID for resin-device-progress patch
VERSION has a format of eg. a.b.c+revX, while VERSION_ID is
a.b.c (at least for non RC, non beta releases that we have now).
This makes a better comparison for versions, as previously e.g.
`version_gt "2.6.0+rev1" "2.6.0"` was true, which is not what we
wanted.

Change-type: patch