upgrade-2.x: possibility of redirecting complete script output to file
Previously the entries called with "log ..." were part of the logfile
created (defined by $LOGFILE), but that was missing the output (both
normal and errors) for the commands called in the script.

Doing these changes, all the output will be put in the logfile on the
device, with the downside, that over SSH there's no copy. To receive logs
over SSH, have to call with --no-log (then no logfile, but log on
standar output). This is different behaviour than before, but should
improve logs for when the scripts are called over the proxy (when
only the on-device logs matter)

Change-type: patch