upgrade-2.x: update APIKEY getting logic
There are some devices that have both apiKey and deviceApiKey set,
that is because they are running with a supervisor that cannot
use the deviceApiKey just yet. Borrow the default logic from the
update-resin-supervisor script.
Also combine reads to have a single file read to get all variables.

Change-type: patch