upgrade-1.x-to-2.x: set to be deleted partitions to read only and flush buffer
In some rare, and hard-to-reproduce cases, when parted removes a partition,
the partition table information fails to update, because maybe
the partition is somehow still in use even if it was unmounted successfully.

Trying to prevent that case by setting the soon-to-be-removed partitions
to read-only and flushing the device buffers, before removing the partition.

Change-type: patch