upgrade-1.x-to-2.x: check if btrfs partition was successfully umounted
In some circumstances, it seems like there are lingering processes
after the data partition, which is btrfs on resinOS 1.x, is umounted.
When that's the case, continuing to the next step is not desirable,
as the following steps are trying to use the partition that is still
in use, and will fail, resulting in a broken-modified device file system.

In these cases, manual resinHUP is needed, most likely, to update the
device. So here just bail out before anything is changed in the first test.
Then the device can be manually rebooted, and will come online as was before.

The second test is less likely to fail (as it's a newly created btrfs partition),
but it might, so better test there too.

Change-type: patch