upgrade-2.x: new flags to pull resinOS from custom repository
Introducing a couple of new flags:

* --resinos-registry: the docker registry to pull from, defaults to Docker Hub
* --resinos-repo: the docker repository in that registry, defaults to resin/resinos
* --resinos-tag: overriding the automatic inference of the resinOS image tag
* --force-slug: override automatic slug detection of the script

It also modifies the behaviour of the --staging flag. That flag was
only used to infer the resinos repo, so it now acts as if
--resinos-repo resin/resinos-staging was set.

Change unrecognized options to warnings for the update scripts. This
is to enable in the same time using these new flags on the resinOS 2.x->2.x
updates, while not needing to implement them in the 1.x->2.x scripts,
which are in maintenance mode.

Change-type: minor