upgrade-2.x: upgrading a `.dev` device switched from error to warning
Treating that update attempt as an error is likely not really helpful.
Since the resin.io dashboard does not enable updating `.dev` devices
anyways, but in support we do some updates for dev device updates
(which is then intentional and cleared with the users), if we just
warn in the logs should be enough for the time being, until a better
story is developed with .dev updates in general.

Change-type: minor