fix cleaning up resinOS update package
For newer supervisor versions (7.x.y) the supervisor does not clean
up images it doesn't know of. We had taken a step before to clean
up the image after update, but we've cleaned it up too early.
The result was that the supervisor update had to repull the os
image.

Now push the cleanup to the final step when it can safely
be removed, as not needed for any other procedures.

Change-type: patch