when first umounting file systems, also stop the relevant systemd services
If services are not stopped, systemd might remount them at awkward times.
This mostly affects `/mnt/data` and its various mounts.
It's also required only the first time when umounting to stop the services.
Further in the process since partitions are manually mounted, there's no
services to stop.

Also, do not stop the service for `mnt-boot.mount` nor `mnt-conf.mount`,
as that will kill `resin-openvpn.service`, and it's not really reqired
for the following steps.

Change-type: patch
Signed-off-by: Gergely Imreh <imrehg@gmail.com>