upgrade-1.x-to-2.x: force dd to physically write stuff to disk before finishing
This might help with some SD cards that are slow to write to disk. Might not,
if the card's faulty or tells the system that things are written to disk, but
actually not. Still should be a general improvement upon the previous state,
even if `sync` was already called by the code later, just before reboot.

Change-type: patch