update-2.x: workaround for RPi1 2.x devices not being able do some docker pulls
For Raspberry Pi 1 devices with docker 17.03 (resinOS 2.0.0-2.5.1) are
not able to currently pull multilayer images from Docker Hub, such
as supervisor images.

Modify the docker command line to limit concurrent downloads, which
seems to be a workaround to this issue. Do the workaround both
for production and development images as needed.

Change-type: patch