upgrade-2.x: Only switch docker to single download where available
For an ongoing issue with ARMv6 devices downloading multilayer images
from Docker HUB, we need to switch docker to not run parallel
downloads. The option used for that is only available for resinOS
versions 2.0.7+ and versions below 2.5.0 are affeccted. Thus only
cover those versions with the immediate fix, which is mainly needed
for the supervisor update only.

For `raspberry-pi` device type (the only one affected) on resinOS
versions <2.0.7, the supervisor update will not succeed in the pull
step, but the device will self-recover ~15 minutes after reboot, when
repulling the supervisor by the regular scheduled supervisor updater.
This inconvenience is a tradeoff, as we don't know any other workaround
for this issue yet.

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@resin.io>