upgrade-1.x: Properly handle updates to logstream 1.x versions
We have some resinOS 1.x versions that have backported, logstream
enabled supervisors. The updater checks the `VERSION` file online
to see what supervisor version to update to. Previously the version
in that file was `a.b.c-shortsha`, and we've thrown away anything
after the dash. In this case we have to communicate that it's a
logstream supervisor, and the cleanest is storing `a.b.c-logstream`
in that file. The updater with this change turns that version into
a supervisor tag, `va.b.c_logstream` ultimately, while backwards
compatible with non-logstream versions online (as long as they
are available).

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@balena.io>