upgrade-2.x: Remove supervisor container not just stop
When stopping the supervisor service and container,
also remove the container, so next time when it starts
up it will be recreated.

This might be important for updates, as there could be a
race condition, the new supervisor starting up on the old
system, and if the container is just stopped, when rebooting
the restart will carry over strange states.

This possibly fixes issues when a lockfile is detected by
the supervisor, but it might be just from an earlier
run of supervisor that was stopped at the right time.
If no container, no data is carried over.

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@balena.io>