upgrade-1.x-to-2.x: Fix incorrect migration to the state partition
Previously some `etc/{docker,dropbear}` folders were created on the
state partition but the files intended to go there were not copied
to the right place. Fix the path.
Likely this is not fatal, as we've done 1000s of updates so far with
this in place, but the fix will make the migration correct as intended.

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@balena.io>