upgrade-2.x: expand on leftover data clearence after interrupted update
The previous way didn't work with `overlay2` storage drivers, as the
directory structure is different.

Also, for more correctness, restart the `balena-host` service after
the cleanup so it would have proper information after the deletes.
Without this `overlay2` cleanup doesn't work, and should be proper
for the `aufs` drivers too.

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@balena.io>