upgrade-1.x-to-2.x: do not umount the boot partition
There seem to be no need for that partition to be actually
umounted/remounted in the update process.

On the other hand, if there's an issue between those mount steps,
we won't be able to connect to the device (as the `config.json`
won't be accessible, and the proxy connectivity check will not
allow connection, not being able to read the device's UUID).

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@balena.io>