upgrade-2.x: user tempfile for FIFO
Using temp files instead of fixed file name on the host OS
side fixes an issue when the update has been interrupted, but
the OS hasn't been restarted before the next update attempt. In that
case repeated `mkfifo` would fail with the same filename. Instead
of reusing the filename, properly use a tempfile that is not
clashing. Thus if the update fails, it can be safely restarted
(and a new filename would be used).

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@balena.io>