upgrade-1.x-to-2.x: unify data partition unmounting and add preventing umount step
Make data partition unmount a reusable function, and add an extra umount step
for one of the partition creation, which seemed to have triggered a remount
in some old SD card cases. This umount step was verified to fix things
with the affected devices, when done in a manual update.

Connects-to: #229
Change-type: patch
Signed-off-by: Gergely Imreh <gergely@balena.io>