upgrade-2.x: increase timeout values during update to 5 minutes
The previous timeout was 0.2s*150=30s, which is got hit by some otherwise
normally working devices, that might be a) slower than usual, b) the user
application might be taking too much of the resources.
Increase this x10, 0.2s*1500=300s=5min, which is still quite short, and
definitely shorter than if the user needs to ask support and retry because
of "spurious" timeouts. Valid errors should be still caught quite quickly.

Change-type: patch
Signed-off-by: Gergely Imreh <gergely@balena.io>