Remove "--nolog" flag from 2.x updater, and always set up logfile
Previously we might have logged when there wasn't any logfile setup,
for example if any of the argument parsing was throwing an error,
and on the 2.x updater there wasn't any way to directly know what
the error was, just that the update didn't start.

Moving the logfile setup before any logging takes place eliminates
this issue. The flag was also only really used in manual HUP runs
(no way to use in self-serve updates), and it's more relevant to
1.x updates than anything since.

Change-type: minor
Signed-off-by: Gergely Imreh <gergely@balena.io>