supervisor: if available, update supervisor without affecting it's state
Recently the supervisor added a codepath that assumes no files underneath it will change during runtime.
OS update hooks can trigger a condition whereby the supervisor reboots the device during a HUP,
which in turn bricks the device. As a workaround, we can use a transient
unit to ensure that the remote-executed update does not occur alongside
the locally-executed update.

Depends-on: https://github.com/balena-os/meta-balena/pull/1973
Change-type: patch
Signed-off-by: Matthew McGinn <matthew@balena.io>