Remove device sanity checks, as they are no longer needed
The device sanity checks examined two on-device files, in an attempt
to prevent HUPs from installing OS versions incompatible with the
underlying hardware.  It is no longer a useful check, as the script is
normally invoked directly from the dashboard.

The immediate prompt for this change is to enable hands-off
device-type changes for customers with private device types; by
removing the check, we can patch the device type in the API, and leave
OS upgrades to the customer.

Change-type: patch
Signed-off-by: Hugh Brown <hugh@balena.io>