private DTs: do not log out from the registry until cleanup
since older OSes get a sideloaded balenaOS hostapp, it is untagged in
the local image store. when we crack that image open to find our new
target supervisor, we have to contact the upstream registry and retag
our local image with the registry reference. in order to do that, we
must remain logged in.

Connects-to: #331
Change-type: patch
Signed-off-by: Matthew McGinn <matthew@balena.io>