upgrade-2.x: report update failure if supervisor state cannot be updated
Currently the supervisor is only updated after the hostapp has already
done so. A failure to patch the target state with the supervisor version
would leave a mismatch between hostOS and supervisor version which for
ESR releases cannot be solved by a supervisor self-update.

Reporting a failed update if the supervisor cannot be set in the target
state allows for a retry before a reboot, avoiding the mismatch. If the
device is rebooted without a successful update, the mismatch still exists
as it does today.

A complete resolution of this problem requires the implementation of
supervisor driven updates as the currently HUP design does not allow to
implement this improvement.

Change-type:patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>