upgrade-2.x.sh: use POSIX redirects
The script redirects stdout to a log file, and stderr both to a log file
and to the console. It does so using bash process substituon, which does
not inherint signal handling from the parent process.

So, by redirecting stdout as:
```
exec 2> >(tee -a "$LOGFILE" >&2)
```

The signals that are trapped by the main process are not trapped by
the spawned processes (tee in this case), and it results on the spawned
process dying, the pipe being broken, and the main process exiting as
it can no longer write to stderr, without the corresponding signal
handlers being run.

This explains the support reported issue of the HUP lock file
being left around, as the trap EXIT handler is not run.

It also allows for the INT ERR TERM handler to work as expected and should
result on hostOS updates to report as failed and not progressing when
the process is interrupted.

Change-type:patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>