upgrade-2.x: Continue update if supervisor version is higher than target
Before patching the supervisor version in the API, the script checks that
the existing version is lower than the target version.

However, there can be a race condition between the check and the actual
patch operation where a self-serve supervisor update could patch the
target version. In that case, the patch operation returns a bad request
(400) code that is indistinguishable from other bad request errors.

The script could parse the request output, but that is bad practise and
likely to break in future. Instead, when a 400 is received from the patch
operation the script checks that the API supervisor version is still lower
than the target version before failing the update. If the target supervisor
version happens to be higher than the target version the update is
allowed to continue as that is a valid scenario.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>