upgrade-2.x.sh: Avoid logging duplicates in transient systemd units
When a transient systemd unit calls into a hostOS script that uses the
os-helpers logging functions, the script already logs to systemd
and stderr.

If the script is then run on a systemd unit that also outputs to stderr,
messages are duplicated in the journal. This is prevented by specifying
to only append stderr output to the log file.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>